New Tag v7-6-3.3

+++ bin/hadd.sh
@@ -1,4 +1,6 @@
 #! /bin/sh
 
-rm $1$2
+if [ -f $1$2 ]; then
+    rm $1$2
+fi
 hadd $1$2 $3
